+++ app/src/main/res/layout/fragment_home.xml
@@ -14,7 +14,7 @@
 
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".views.MainActivity">
 
         <TextView
             android:id="@+id/about_tv"
